Creating MemberSearch component that checks valid member data for matchChecker
