+++ frontend/src/components/Admin/EditTeam.tsx
@@ -4,7 +4,7 @@ import styles from './EditTeam.module.css';
 import ErrorModal from '../Modals/ErrorModal';
 import { APICache, Emitters } from '../../utils';
 import { Team, TeamsAPI } from '../../API/TeamsAPI';
-import CustomSearch from '../Common/Search';
+import CustomSearch, { MemberSearch } from '../Common/Search';
 import { MembersAPI, Member } from '../../API/MembersAPI';
 
 type EditTeamState = {
@@ -216,18 +216,9 @@ class EditTeam extends React.Component<Record<string, unknown>, EditTeamState> {
                             <Label>{mem.email}</Label>
                           </Segment>
                         )}
-                        matchChecker={(query: string, member: Member) => {
-                          const queryLower = query.toLowerCase();
-                          return (
-                            member.email.toLowerCase().startsWith(queryLower) ||
-                            member.firstName.toLowerCase().startsWith(queryLower) ||
-                            member.lastName.toLowerCase().startsWith(queryLower) ||
-                            member.role.toLowerCase().startsWith(queryLower) ||
-                            `${member.firstName.toLowerCase()} ${member.lastName.toLowerCase()}`.startsWith(
-                              queryLower
-                            )
-                          );
-                        }}
+                        matchChecker={(query: string, member: Member) =>
+                          MemberSearch(query, member)
+                        }
                         selectCallback={(mem: Member) => {
                           if (this.state.currentSelectedTeam) {
                             this.setState({
@@ -287,18 +278,9 @@ class EditTeam extends React.Component<Record<string, unknown>, EditTeamState> {
                             <Label>{mem.email}</Label>
                           </Segment>
                         )}
-                        matchChecker={(query: string, member: Member) => {
-                          const queryLower = query.toLowerCase();
-                          return (
-                            member.email.toLowerCase().startsWith(queryLower) ||
-                            member.firstName.toLowerCase().startsWith(queryLower) ||
-                            member.lastName.toLowerCase().startsWith(queryLower) ||
-                            member.role.toLowerCase().startsWith(queryLower) ||
-                            `${member.firstName.toLowerCase()} ${member.lastName.toLowerCase()}`.startsWith(
-                              queryLower
-                            )
-                          );
-                        }}
+                        matchChecker={(query: string, member: Member) =>
+                          MemberSearch(query, member)
+                        }
                         selectCallback={(mem: Member) => {
                           if (this.state.currentSelectedTeam) {
                             this.setState({
@@ -358,18 +340,9 @@ class EditTeam extends React.Component<Record<string, unknown>, EditTeamState> {
                             <Label>{mem.email}</Label>
                           </Segment>
                         )}
-                        matchChecker={(query: string, member: Member) => {
-                          const queryLower = query.toLowerCase();
-                          return (
-                            member.email.toLowerCase().startsWith(queryLower) ||
-                            member.firstName.toLowerCase().startsWith(queryLower) ||
-                            member.lastName.toLowerCase().startsWith(queryLower) ||
-                            member.role.toLowerCase().startsWith(queryLower) ||
-                            `${member.firstName.toLowerCase()} ${member.lastName.toLowerCase()}`.startsWith(
-                              queryLower
-                            )
-                          );
-                        }}
+                        matchChecker={(query: string, member: Member) =>
+                          MemberSearch(query, member)
+                        }
                         selectCallback={(mem: Member) => {
                           if (this.state.currentSelectedTeam) {
                             this.setState({
